+++ src/command/owner-commands/live.ts
@@ -1,6 +1,14 @@
 import { Platform } from '../../enums';
 import { ENV } from '../../config/config.service';
 
+/**
+ * The !live command without any arguments will announce that the streamer is live on discord if
+ * discord is enabled. It will create a message using AI if no message is provided.
+ *
+ * If the body of the command contains "nodiscord" then the discord announcement will be skipped. This
+ * is sometimes necessary if the bot crashed and you want to go back into live mode without
+ * announcing it on discord.
+ */
 const command: Command = {
   name: 'live',
   ownerOnly: true,
@@ -11,22 +19,21 @@ const command: Command = {
     commandState.isLive = true;
     await services.twitchService.ownerRunCommand('!autochat on');
 
-    if (!ENV.DISCORD_ENABLED) {
+    if (!ENV.DISCORD_ENABLED || msg.toLowerCase().includes('nodiscord')) {
       return true;
     }
 
     if (!msg) {
       try {
         msg = await services.openaiService.sendPrompt(
-          'The blunder master is about to go live on twitch. Say something to get people excited',
+          `${ENV.NICKNAME} is about to go live on twitch. Say something to get people excited`,
           {
             temp: 1.4,
             includeBlunderBotContext: true
           }
         );
       } catch (e) {
-        msg =
-          "Are you ready? It's time. The Blunder Master is here and he's ready to play some chess!";
+        msg = `Are you ready? It's time. ${ENV.NICKNAME} is here and he's ready to play some chess!`;
       }
     }